Add compatibility with Rails 6 and Zeitwerk
- Move decorators to app/decorators in order to adapt the gem to the rails 6
- Remove lib from the autoload path for Zeitwerk compatibility
