+++ lib/solidus_sale_prices/engine.rb
@@ -2,8 +2,6 @@ module SolidusSalePrices
   class Engine < Rails::Engine
     engine_name 'solidus_sale_prices'
 
-    config.autoload_paths += %W(#{config.root}/lib)
-
     # use rspec for tests
     config.generators do |g|
       g.test_framework :rspec
